Implement notification channel
The notification channel is for all communication between the
bnet server and the game server.

Pegasus Util packet handling has been reworked to use the
PacketID enums in the protos.

A goroutine leak has been stopped, and facilities for any
goroutine to stop itself once the client has disconnected have
been added.

Many error conditions have been reworked to simply panic
instead of returning an error as an argument.  I don't
foresee needing to check the error argument for any of these
cases in the future, as usually the desired behavior is to
disconnect the client in any error condition.

A single test has found its way in to the tree.  Rejoice!